+++ api/specs/web-server/_projects_groups.py
@@ -10,11 +10,11 @@ from fastapi import APIRouter, Depends, status
 from models_library.generics import Envelope
 from simcore_service_webserver._meta import API_VTAG
 from simcore_service_webserver.projects._common.models import ProjectPathParams
-from simcore_service_webserver.projects._groups_api import ProjectGroupGet
-from simcore_service_webserver.projects._groups_rest import (
+from simcore_service_webserver.projects._controller._groups_rest import (
     _ProjectsGroupsBodyParams,
     _ProjectsGroupsPathParams,
 )
+from simcore_service_webserver.projects._groups_api import ProjectGroupGet
 
 router = APIRouter(
     prefix=f"/{API_VTAG}",
