+++ api/specs/web-server/_projects_metadata.py
@@ -18,7 +18,9 @@ from models_library.api_schemas_webserver.projects_metadata import (
 )
 from models_library.generics import Envelope
 from simcore_service_webserver._meta import API_VTAG
-from simcore_service_webserver.projects._metadata_rest import ProjectPathParams
+from simcore_service_webserver.projects._controller._metadata_rest import (
+    ProjectPathParams,
+)
 
 router = APIRouter(prefix=f"/{API_VTAG}", tags=["projects", "metadata"])
