+++ api/specs/web-server/_projects_states.py
@@ -14,7 +14,7 @@ from pydantic import ValidationError
 from servicelib.aiohttp import status
 from simcore_service_webserver._meta import API_VTAG
 from simcore_service_webserver.director_v2.exceptions import DirectorServiceError
-from simcore_service_webserver.projects._states_handlers import (
+from simcore_service_webserver.projects._controller._states_handlers import (
     ProjectPathParams,
     _OpenProjectQuery,
 )
@@ -70,20 +70,17 @@ def open_project(
     client_session_id: Annotated[str, Body(...)],
     _path_params: Annotated[ProjectPathParams, Depends()],
     _query_params: Annotated[_OpenProjectQuery, Depends()],
-):
-    ...
+): ...
 
 
 @router.post("/projects/{project_id}:close", status_code=status.HTTP_204_NO_CONTENT)
 def close_project(
     _path_params: Annotated[ProjectPathParams, Depends()],
     client_session_id: Annotated[str, Body(...)],
-):
-    ...
+): ...
 
 
 @router.get("/projects/{project_id}/state", response_model=Envelope[ProjectState])
 def get_project_state(
     _path_params: Annotated[ProjectPathParams, Depends()],
-):
-    ...
+): ...
