+++ services/web/server/src/simcore_service_webserver/projects/_controller/_nodes_rest.py
@@ -56,23 +56,23 @@ from servicelib.rabbitmq.rpc_interfaces.dynamic_scheduler.errors import (
 from servicelib.services_utils import get_status_as_dict
 from simcore_postgres_database.models.users import UserRole
 
-from .._meta import API_VTAG as VTAG
-from ..catalog import catalog_service
-from ..dynamic_scheduler import api as dynamic_scheduler_api
-from ..groups.api import get_group_from_gid, list_all_user_groups_ids
-from ..groups.exceptions import GroupNotFoundError
-from ..login.decorators import login_required
-from ..security.decorators import permission_required
-from ..users.api import get_user_id_from_gid, get_user_role
-from ..utils_aiohttp import envelope_json_response
-from . import _access_rights_api as access_rights_service
-from . import _nodes_api as _nodes_service
-from . import nodes_utils, projects_service
-from ._common.exceptions_handlers import handle_plugin_requests_exceptions
-from ._common.models import ProjectPathParams, RequestContext
-from ._nodes_api import NodeScreenshot, get_node_screenshots
-from .api import has_user_project_access_rights
-from .exceptions import (
+from ..._meta import API_VTAG as VTAG
+from ...catalog import catalog_service
+from ...dynamic_scheduler import api as dynamic_scheduler_api
+from ...groups.api import get_group_from_gid, list_all_user_groups_ids
+from ...groups.exceptions import GroupNotFoundError
+from ...login.decorators import login_required
+from ...security.decorators import permission_required
+from ...users.api import get_user_id_from_gid, get_user_role
+from ...utils_aiohttp import envelope_json_response
+from .. import _access_rights_api as access_rights_service
+from .. import _nodes_api as _nodes_service
+from .. import nodes_utils, projects_service
+from .._common.exceptions_handlers import handle_plugin_requests_exceptions
+from .._common.models import ProjectPathParams, RequestContext
+from .._nodes_api import NodeScreenshot, get_node_screenshots
+from ..api import has_user_project_access_rights
+from ..exceptions import (
     NodeNotFoundError,
     ProjectNodeResourcesInsufficientRightsError,
     ProjectNodeResourcesInvalidError,
