+++ services/web/server/src/simcore_service_webserver/projects/_controller/_ports_rest.py
@@ -25,16 +25,16 @@ from servicelib.aiohttp.requests_validation import (
     parse_request_path_parameters_as,
 )
 
-from .._meta import API_VTAG as VTAG
-from ..login.decorators import login_required
-from ..security.decorators import permission_required
-from ..utils_aiohttp import envelope_json_response
-from . import _ports_api, projects_service
-from ._access_rights_api import check_user_project_permission
-from ._common.exceptions_handlers import handle_plugin_requests_exceptions
-from ._common.models import ProjectPathParams, RequestContext
-from .db import ProjectDBAPI
-from .models import ProjectDict
+from ..._meta import API_VTAG as VTAG
+from ...login.decorators import login_required
+from ...security.decorators import permission_required
+from ...utils_aiohttp import envelope_json_response
+from .. import _ports_api, projects_service
+from .._access_rights_api import check_user_project_permission
+from .._common.exceptions_handlers import handle_plugin_requests_exceptions
+from .._common.models import ProjectPathParams, RequestContext
+from ..db import ProjectDBAPI
+from ..models import ProjectDict
 
 log = logging.getLogger(__name__)
