+++ services/web/server/src/simcore_service_webserver/projects/_controller/_trash_rest.py
